moved AdditionalCredentials back to SubversionSCM
