+++ src/test/java/hudson/scm/CredentialsExternalsTest.java
@@ -116,7 +116,7 @@ public class CredentialsExternalsTest {
                 p.setScm(new SubversionSCM(
                     Collections.singletonList(new SubversionSCM.ModuleLocation("svn://localhost:" + mainPort + "/prj/trunk", "main-creds", ".", "", false)),
                     null, null, null, null, null, null, null, false, false, // WTF was all that?
-                    Collections.singletonList(new AdditionalCredentials("<svn://localhost:" + extPort + "> " + ext.uuid(), "ext-creds"))));
+                    Collections.singletonList(new SubversionSCM.AdditionalCredentials("<svn://localhost:" + extPort + "> " + ext.uuid(), "ext-creds"))));
                 FreeStyleBuild b = r.buildAndAssertSuccess(p);
                 assertEquals("", b.getWorkspace().child("file").readToString());
                 assertEquals("", b.getWorkspace().child("ext/file").readToString());
